+++ src/main/kotlin/racingcar/Application.kt
@@ -1,22 +1,8 @@
 package racingcar
 
-import racingcar.model.Car
-import racingcar.model.CarMove
-import racingcar.model.RandomGenerator
-import racingcar.view.InputView
-import racingcar.view.OutputView
+import racingcar.controller.RacingController
 
 fun main() {
-    val inputView = InputView()
-    val outputView = OutputView()
-
-    val carNames = inputView.getCarNames()
-    val roundCount = inputView.getRoundCount()
-
-    val cars = carNames.map { Car(it, CarMove(RandomGenerator())) }
-
-    repeat(roundCount) {
-        cars.forEach { it.move() }
-        outputView.printRoundResult(cars)
-    }
+    val racingController = RacingController()
+    racingController.start()
 }
