+++ src/main/kotlin/racingcar/view/OutputView.kt
@@ -7,4 +7,12 @@ class OutputView {
         cars.forEach { println(it) }
         println()
     }
+
+    fun gameResult() {
+        println(GAME_RESULT_MESSAGE)
+    }
+
+    companion object {
+        const val GAME_RESULT_MESSAGE = "실행 결과"
+    }
 }
